Cache proto objects per series to reduce allocations
So far we generated the entire series proto objects on the fly for each
sample. This was very CPU intensive due to the allocations it did and
garbage it produced.
This creates the proto objects when series records are read from the WAL
and caches them. This reduces CPU usage by 10x and increases the maximum
sample throughput by 10x.

Signed-off-by: Fabian Reinartz <freinartz@google.com>