Add option to store payload to local files.
This PR introduces an option to send the payload to local files
inside temporary directories. It is useful for troubleshooting if the
data cannot be sent to Monitoring API.

This PR also introduces a new type of client: file client. It will
create a temporary file, and write CreateTimeSeriesRequest to the file.