Publish standard grpc.io metrics.
We were publishing a different version with an extra label, but had kept the upstream name of the metrics. This makes querying this metric across different components harder (or impossible).

This drops the grpc_client_status label from most metrics, but it shouldn't be a big loss, since metrics like request size are rarely correlated with an error.